feat: Created MQ procedure for computing from 3 matrices
