+++ math/mq.go
@@ -10,6 +10,51 @@ func MQ(F []uint8, x []uint8, m int) []uint8 {
 	return fx
 }
 
+func MQUOV(P1i, P2i, P3i, x []uint8, m int) []uint8 {
+	fx := make([]uint8, 0)
+	n := len(x)
+
+	vec := NewVector(x)
+
+	lenP1 := (n - m) * (n - m + 1) / 2
+	lenP2 := (n - m) * m
+	lenP3 := m * (m + 1) / 2
+
+	for k := 0; k < m; k++ {
+		var acc uint8 = 0
+		P1 := NewUpperTriangle(
+			NewDenseMatrix(
+				n-m, n-m,
+				P1i[k*lenP1:(k+1)*lenP1],
+			),
+		)
+		P2 := NewDenseMatrix(
+			n-m, m, P2i[k*lenP2:(k+1)*lenP2],
+		)
+		P3 := NewUpperTriangle(
+			NewDenseMatrix(
+				m, m, P3i[k*lenP3:(k+1)*lenP3],
+			),
+		)
+		for i := 0; i < n; i++ {
+			for j := i; j < n; j++ {
+				t := Mul(vec.At(i, 0), vec.At(j, 0))
+				if j < n-m {
+					acc ^= Mul(t, P1.At(i, j))
+				} else {
+					if i < n-m {
+						acc ^= Mul(t, P2.At(i, j-n+m))
+					} else {
+						acc ^= Mul(t, P3.At(i-n+m, j-n+m))
+					}
+				}
+			}
+		}
+		fx = append(fx, acc)
+	}
+	return fx
+}
+
 func G(F []uint8, x []uint8, y []uint8, m int) []uint8 {
 	var n int
 	if n = len(x); n != len(y) {
